Move tket2 extension loading and make bool_type_def opaque
