+++ guppylang/module.py
@@ -362,7 +362,7 @@ class GuppyModule:
         # compute this dynamically from the imported dependencies instead.
         #
         # The hugr prelude and std_extensions are implicit.
-        from guppylang.std._internal.compiler.quantum import TKET2_EXTENSIONS
+        from guppylang.std._internal.compiler.tket2_exts import TKET2_EXTENSIONS
 
         extensions = [*TKET2_EXTENSIONS, guppylang.compiler.hugr_extension.EXTENSION]
