+++ guppylang/std/_internal/compiler/quantum.py
@@ -8,46 +8,23 @@ from hugr import Wire, ops
 from hugr import ext as he
 from hugr import tys as ht
 from hugr.std.float import FLOAT_T
-from tket2_exts import (
-    futures,
-    qsystem,
-    qsystem_random,
-    qsystem_utils,
-    quantum,
-    result,
-    rotation,
-)
+
 
 from guppylang.definition.custom import CustomInoutCallCompiler
 from guppylang.definition.value import CallReturnWires
+from guppylang.std._internal.compiler.tket2_exts import QSYSTEM_RANDOM_EXTENSION, QUANTUM_EXTENSION, ROTATION_EXTENSION
 
 # ----------------------------------------------
 # --------- tket2.* extensions -----------------
 # ----------------------------------------------
 
-FUTURES_EXTENSION = futures()
-QSYSTEM_EXTENSION = qsystem()
-QSYSTEM_RANDOM_EXTENSION = qsystem_random()
-QSYSTEM_UTILS_EXTENSION = qsystem_utils()
-QUANTUM_EXTENSION = quantum()
-RESULT_EXTENSION = result()
-ROTATION_EXTENSION = rotation()
+
 
 RNGCONTEXT_T_DEF = QSYSTEM_RANDOM_EXTENSION.get_type("context")
 RNGCONTEXT_T = ht.ExtType(RNGCONTEXT_T_DEF)
 
 ROTATION_T_DEF = ROTATION_EXTENSION.get_type("rotation")
 ROTATION_T = ht.ExtType(ROTATION_T_DEF)
-
-TKET2_EXTENSIONS = [
-    FUTURES_EXTENSION,
-    QSYSTEM_EXTENSION,
-    QSYSTEM_RANDOM_EXTENSION,
-    QSYSTEM_UTILS_EXTENSION,
-    QUANTUM_EXTENSION,
-    RESULT_EXTENSION,
-    ROTATION_EXTENSION,
-]
 
 
 def from_halfturns_unchecked() -> ops.ExtOp:
