+++ guppylang/std/qsystem/utils.py
@@ -2,7 +2,7 @@ from typing import no_type_check
 
 from guppylang.decorator import guppy
 from guppylang.module import GuppyModule
-from guppylang.std._internal.compiler.quantum import QSYSTEM_UTILS_EXTENSION
+from guppylang.std._internal.compiler.tket2_exts import QSYSTEM_UTILS_EXTENSION
 from guppylang.std._internal.util import external_op
 
 qsystem_utils = GuppyModule("qsystem.utils")
