+++ tests/integration/test_side_effect_ordering.py
@@ -7,7 +7,7 @@ from hugr.std import PRELUDE
 
 from guppylang import guppy
 from guppylang.module import GuppyModule
-from guppylang.std._internal.compiler.quantum import RESULT_EXTENSION, QUANTUM_EXTENSION
+from guppylang.std._internal.compiler.tket2_exts import RESULT_EXTENSION, QUANTUM_EXTENSION
 from guppylang.std.builtins import result, array, owned, panic
 from guppylang.std.quantum import qubit, discard, measure
